+++ src/main/ruby/truffleruby/core/thread.rb
@@ -346,6 +346,18 @@ class Thread::Backtrace
 end
 
 class Thread::Backtrace::Location
+  def syntax_tree
+    require 'prism'
+    range = self.source_range
+    return nil unless range&.absolute_path
+    result = Prism.parse_file(range.absolute_path, raise_error: true)
+    start_offset = result.source.byte_offset(range.start_line, range.start_column)
+    end_offset = result.source.byte_offset(range.end_line, range.end_column)
+    result.value.tunnel(range.start_line, range.start_column).rfind do |n|
+      n.start_offset == start_offset && n.end_offset == end_offset
+    end
+  end
+
   def inspect
     to_s.inspect
   end